# LibGfx/JPEG: Performance work regarding the Huffman stream