If this PR is too much to review in a single shot, I can easily split it!

Results of `BenchmarkJPEGLoader`:

<table>
<tr>
 <td>
 <td>small_image
 <td>big_image
 <td>rgb_image
 <td>several_scans
<tr>
 <td>Master
 <td>0.0±0.0ms
 <td>601.7±4.7ms
 <td>9.0±0.0ms
 <td>7.9±0.3ms
<tr>
 <td>Using `peek` in `read`
 <td>0.0±0.0ms
 <td>557.7±32.3ms
 <td>8.1±0.6ms
 <td>7.9±0.3ms
<tr>
 <td>Using `peek` in `next_symbol`
 <td>0.0±0.0ms
 <td>526.1±40.6ms
 <td>8.0±0.3ms
 <td>7.7±0.6ms
<tr>
 <td>Using the Lookup table
 <td>0.0±0.0ms
 <td>459.8±34ms
 <td>7.9±0.3ms
 <td>7.2±0.4ms
</table>

With this image:
https://commons.wikimedia.org/wiki/File:Seven_Coloured_Pencils_(sRGB).jpg
`image` is now only 3x (previously 4) slower than libjpeg-turbo's `djpeg`!